mingw: fix signed/unsigned comparison warning
DWORD is an unsigned long, but it's assigned -1 in various places.
mingw warns when comparing a DWORD value against -1. This commit
casts the -1 to DWORD to avoid the warning.

+++ vdagent/desktop_layout.cpp
@@ -263,7 +263,7 @@ bool DesktopLayout::init_dev_mode(LPCTSTR dev_name, DEVMODE* dev_mode, DisplayMo
             best = i;
         }
     }
-    if (best == -1 || !EnumDisplaySettings(dev_name, best, dev_mode)) {
+    if (best == (DWORD)-1 || !EnumDisplaySettings(dev_name, best, dev_mode)) {
         return false;
     }
     dev_mode->dmFields = DM_BITSPERPEL | DM_PELSWIDTH | DM_PELSHEIGHT;
